requirements.txt
openai==2.8.0
pydantic==2.12.4
pydantic_settings==2.12.0
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5